requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albucore=0.0.23=pypi_0
albumentations=2.0.5=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
beautifulsoup4=4.13.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
cython=3.0.12=pypi_0
debugpy=1.8.11=py312h6a678d5_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h6a678d5_0
filelock=3.13.1=py312h06a4308_0
fire=0.7.0=pypi_0
fonttools=4.56.0=pypi_0
fsspec=2024.12.0=py312h06a4308_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py312h5eee18b_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=9.0.2=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=py312h06a4308_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=4.25.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lmdb=1.6.2=pypi_0
lxml=5.3.1=pypi_0
markupsafe=3.0.2=py312h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_2
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py312h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=py312h06a4308_0
numpy=2.2.4=pypi_0
numpy-base=2.0.1=py312h0da6c21_1
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.16=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
paddleocr=2.10.0=pypi_0
paddlepaddle-gpu=2.6.2=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=25.0.1=pypi_0
platformdirs=4.3.7=pyh29332c3_0
prompt-toolkit=3.0.50=pyha770c72_0
protobuf=6.30.1=pypi_0
psutil=5.9.0=py312h5eee18b_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyclipper=1.3.0.post6=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
python=3.12.9=h5148396_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-docx=1.1.2=pypi_0
pytorch=2.3.1=cpu_py312hece47aa_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py312h6a678d5_0
rapidfuzz=3.12.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-image=0.25.2=pypi_0
scipy=1.15.2=pypi_0
setuptools=75.8.0=py312h06a4308_0
shapely=2.0.7=pypi_0
simsimd=6.2.1=pypi_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
stringzilla=3.12.3=pypi_0
sympy=1.13.3=py312h06a4308_1
tbb=2021.8.0=hdb19cb5_0
termcolor=2.5.0=pypi_0
tifffile=2025.3.13=pypi_0
tk=8.6.14=h39e8969_0
tornado=6.4.2=py312h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py312h06a4308_0
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
